Add almost-complete support for Legion
The appropriate number are added and it's allowed to have one fewer
Legion (which is a judgment call, one more Legion is also legal but
seems unfair). The only oddity is that the shown list of goal
distributions is awkwardly long and unhelpful, since any combination of
townsfolk and outsiders are allowed.