Fetch icons from pocket-grimoire
The script tool switched to official art, so we can't use it any more.
The pocket grimoire already has all the playtest art + unofficial tokens
for unreleased characters, so we use that. Unfortunately it is a 45MB
ZIP file to download, but it's being served by GitHub so load shouldn't
be an issue.