Make sure Drunk is considered base 3
Drunk is exceptional in that it's a base character with an override that
adds a nightsheet ability (to remind the storyteller to assign a Drunk).
It was previously using the homebrew support to give it an index, which
had the minor but unfortunate side effect of making the drunk considered
a homebrew role.

It's more of a special case now which is too bad but seemed an
acceptable compromise. I can see this being used again for Minstrel,
which is similarly easy to forget to check each night.