Use new townsquare data for roles
The new upstream data source is nicholas-eden/townsquare, which is kept
up-to-date fairly rapidly. This change overhauls the data to be
up-to-date and reduces the dependence on overrides to add new roles;
they be automatically incorporated with ability text and night
instructions.

Along the way I also switched to the same source for the nightsheet,
which now uses identifiers rather than names. We no longer use any data
from the official script tool, which also protects against future
incompatibilities if and when those JSON files change.

Fixes #32 (new Townsfolk Acrobat)