Faster searching with Fuse.js (#38)
* Resolves https://github.com/tchajed/botc-tools/issues/15.

Unlike `match-sorter`, Fuse.js creates an index which is cached in
between renders. This drastically cuts down on the time spent writing
queries. We can also debounce so we don't overload the CPU while the
user is typing.

Searching happens in a WebWorker so it doesn't block the UI thread.