Select Minions for Lord of Typhon
With Lord of Typhon, minions do not go in bag. This commit requires them
to be selected (including the +1 Minion effect of Lord of Typhon) but
causes them to be excluded from the bag. This both allows finishing
setup selection for the Storyteller (although they can change the chosen
minions after distributing roles) and puts the in-play Minions on the
nightsheet.

Fixes #21.