Attempt to compress JSON in URL (#73)
Stuffing the JSON in the URL results in a massive URL and a
correspondingly massive QR code that is hard to scan. If we replace
{"id": "character"} with the index of the character in roles.json we
save a lot of space.

Before:


https://botc-tools.xyz/script.html?json=%5B%7B%22id%22%3A%22_meta%22%2C%22name%22%3A%22Magical%20Onion%20Pies%22%7D%2C%7B%22id%22%3A%22noble%22%7D%2C%7B%22id%22%3A%22washerwoman%22%7D%2C%7B%22id%22%3A%22librarian%22%7D%2C%7B%22id%22%3A%22chef%22%7D%2C%7B%22id%22%3A%22pixie%22%7D%2C%7B%22id%22%3A%22empath%22%7D%2C%7B%22id%22%3A%22fortuneteller%22%7D%2C%7B%22id%22%3A%22monk%22%7D%2C%7B%22id%22%3A%22slayer%22%7D%2C%7B%22id%22%3A%22soldier%22%7D%2C%7B%22id%22%3A%22cannibal%22%7D%2C%7B%22id%22%3A%22ravenkeeper%22%7D%2C%7B%22id%22%3A%22mayor%22%7D%2C%7B%22id%22%3A%22virgin%22%7D%2C%7B%22id%22%3A%22ogre%22%7D%2C%7B%22id%22%3A%22drunk%22%7D%2C%7B%22id%22%3A%22recluse%22%7D%2C%7B%22id%22%3A%22saint%22%7D%2C%7B%22id%22%3A%22poisoner%22%7D%2C%7B%22id%22%3A%22wizard%22%7D%2C%7B%22id%22%3A%22marionette%22%7D%2C%7B%22id%22%3A%22scarletwoman%22%7D%2C%7B%22id%22%3A%22baron%22%7D%2C%7B%22id%22%3A%22imp%22%7D%5D

After:


http://localhost:5173/script.html?json=[{%22id%22:%22_meta%22,%22name%22:%22Magical%20Onion%20Pies%22},89,0,1,3,92,4,5,7,10,11,109,8,12,9,118,14,15,16,17,132,135,19,20,21]

I think this is mostly stable given the use case of sharing a QR code
around in a live game but it's not great that an asset update might
break a link if an index points to a new character 😅 open to ideas here
on how to make it a bit more robust to roles.json updates.